Fix water overlays again (#2064)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
Calling cut on the wrong thing.
Removes unused logic from the volume adjustments.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
Oops 600 overlays.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [ ] You tested this on a local server.
- [ ] This code did not runtime during testing.
- [ ] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

+++ code/__HELPERS/icon_smoothing.dm
@@ -216,6 +216,8 @@ DEFINE_BITFIELD(smoothing_junction, list(
 	add_overlay(overlay)
 
 /turf/proc/remove_neighborlays()
+	if(!LAZYLEN(neighborlay_list))
+		return
 	for(var/key as anything in neighborlay_list)
 		cut_overlay(neighborlay_list[key])
 		qdel(neighborlay_list[key])
